+++ josh/server/josh.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"reflect"
 
-	"github.com/deref/exo/jsonutil"
+	"github.com/deref/exo/util/jsonutil"
 )
 
 type MethodHandler struct {
