+++ util/atom/atom.go
